requirements.txt
--no-binary=zxing-cpp
fastapi~=0.101.0
zxing-cpp~=2.1.0
Pillow~=10.0.0
python-multipart~=0.0.6
